fully implemented mentiontype enum and consolidated quote and quoteoptout model files
