+++ models/__init__.py
@@ -5,6 +5,5 @@ from models.messages import *
 from models.models import *
 from models.printing import *
 from models.quote import *
-from models.quoteoptouts import *
 from models.reminder import *
 from models.user import *
