+++ models/quote.py
@@ -14,7 +14,7 @@ from sqlalchemy.orm import relationship
 from models.models import Base, auto_str
 from utils.mentions import MentionType
 
-__all__ = ["Quote"]
+__all__ = ["Quote","QuoteOptouts"]
 
 
 @auto_str
@@ -31,6 +31,15 @@ class Quote(Base):
     author = relationship("User", uselist=False, foreign_keys=author_id)
 
     def author_to_string(self) -> str:
-        if self.author_type == "id":
+        if self.author_type == MentionType.ID:
             return f"<@{self.author.user_uid}>"
         return self.author_string
+
+
+@auto_str
+class QuoteOptouts(Base):
+    __tablename__ = "quotes-opt-out"
+    id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
+    user_type = Column(Enum("id", "string", name="user_type"), nullable=False)
+    user_id = Column(Integer, ForeignKey("users.id"), nullable=True)
+    user_string = Column(String, nullable=True)
